paper-data-extractor/pyproject.toml
[project]
name = "paper-data-extractor"
version = "0.1.0"
description = "Microservice for paper classification from LinkML-backed taxonomy models."
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "jinja2>=3.1.0",
    "python-multipart>=0.0.9",
    "pyyaml>=6.0.0",
    "pydantic>=2.8.0",
    "linkml>=1.8.0",
    "linkml-runtime>=1.8.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "httpx>=0.27.0",
]

[project.scripts]
paper-data-extractor = "paper_data_extractor.main:run"
paper-data-extractor-compile-review-schema = "paper_data_extractor.cli:compile_review_schema_cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/paper_data_extractor"]
